# optimized load_pointcloud_ply by 27.91x (2791% speedup)