--- a/kornia/utils/pointcloud_io.py
+++ b/kornia/utils/pointcloud_io.py
@@ -16,7 +16,7 @@
 #
 
 import os
-
+import numpy as np
 import torch
 
 
@@ -79,24 +79,16 @@ def load_pointcloud_ply(filename: str, header_size: int = 8) -> torch.Tensor:
         :math:`*` represents the number of points.
 
     """
-    if not isinstance(filename, str) and filename[-3:] == ".ply":
-        raise TypeError(f"Input filename must be a string in with the .ply  extension. Got {filename}")
+    if not (isinstance(filename, str) and filename.lower().endswith(".ply")):
+        raise TypeError(f"Input filename must be a string with the .ply extension. Got {filename!r}")
     if not os.path.isfile(filename):
         raise ValueError("Input filename is not an existing file.")
     if not (isinstance(header_size, int) and header_size > 0):
         raise TypeError(f"Input header_size must be a positive integer. Got {header_size}.")
-    # open the file and populate tensor
-    with open(filename) as f:
-        points = []
 
-        # skip header
-        lines = f.readlines()[header_size:]
-
-        # iterate over the points
-        for line in lines:
-            x_str, y_str, z_str = line.split()
-            points.append((torch.tensor(float(x_str)), torch.tensor(float(y_str)), torch.tensor(float(z_str))))
-
-        # create tensor from list
-        pointcloud: torch.Tensor = torch.tensor(points)
-        return pointcloud
+    arr = np.loadtxt(filename, dtype=np.float32, usecols=(0, 1, 2), skiprows=header_size)
+    if arr.ndim == 1:
+        arr = arr.reshape(1, 3)
+    if arr.shape[1] != 3:
+        raise ValueError(f"Expected 3 columns per point, got {arr.shape[1]} columns.")
+    return torch.from_numpy(arr)