optimized load_pointcloud_ply by using numpy load text function to avoid iterating through all points individually


https://colab.research.google.com/drive/1muDCsMq-Kt1b5SLYs3XK6tlz9nnarEyM?usp=sharing
here's a benchmark script to test the validity of the output as well as the speedup provided

--- Testing with 10000 points ---
Validation passed ✅
Original:  0.281743 sec
Optimized: 0.012260 sec
Speedup:   22.98×

--- Testing with 100000 points ---
Validation passed ✅
Original:  3.367795 sec
Optimized: 0.120685 sec
Speedup:   27.91×